+++ Execution/InformaticaPowerCenter/src/com/xbreeze/xtest/process/informaticapowercenter/execution/InformaticaPowerCenterExecutor.java
@@ -69,9 +69,11 @@ public class InformaticaPowerCenterExecutor implements ProcessExecutor {
 	     return ipcs;	     
 	}
 
-
-
-	
-	
-	
+	@Override
+	public void cleanUp() throws XTestProcessException {
+		//Close all existing connections
+		for (InformaticaPowerCenterServerInstance ipcs : _connections.values()) {
+			ipcs.closeConnection();
+		}
+	}
 }