+++ TestCrossTest/src/test/java/com/xbreeze/xtest/test/TestProcessExecutor.java
@@ -19,4 +19,10 @@ public class TestProcessExecutor implements ProcessExecutor {
 		}
 	}
 
+	@Override
+	public void cleanUp() throws XTestProcessException {
+		logger.info("Cleanup invoked");
+		
+	}
+
 }